# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: easyinterview API
  version: 1.0.0
  summary: HTTP contract for easyinterview P0 — JD onboarding, mock interview, evidence-based review.
  description: |
    Single source of truth for all easyinterview HTTP endpoints. Owned by spec
    `docs/spec/openapi-v1-contract/spec.md` (B2). Any change must follow the
    additive-only / breaking-change rules locked in spec §4.4 and ADR review flow.

    Auth path follows ADR-Q1: first-party session cookie minted by
    `GET /api/v1/auth/email/verify` after the magic-link challenge issued by
    `POST /api/v1/auth/email/start`. `Authorization: Bearer` is **not** a P0
    default scheme.
  contact:
    name: easyinterview engineering
    url: https://github.com/monshunter/easyinterview
servers:
  - url: /api/v1
    description: API v1 base path. All endpoints in this document are mounted under this prefix.
...
security:
  - sessionCookie: []
paths:
...
  /profiles/me:
    get:
      tags: [Profile]
      operationId: getMyProfile
      summary: Get my candidate profile
      parameters:
        - $ref: '#/components/parameters/XRequestID'
        - $ref: '#/components/parameters/Traceparent'
        - $ref: '#/components/parameters/AcceptLanguage'
        - $ref: '#/components/parameters/XClientVersion'
      responses:
        '200':
          description: Candidate profile.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/XRequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CandidateProfile'
        default:
          $ref: '#/components/responses/ApiErrorResponse'
    patch:
      tags: [Profile]
      operationId: updateMyProfile
      summary: Update my candidate profile (Lite)
      parameters:
        - $ref: '#/components/parameters/XRequestID'
        - $ref: '#/components/parameters/Traceparent'
        - $ref: '#/components/parameters/AcceptLanguage'
        - $ref: '#/components/parameters/XClientVersion'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
      responses:
        '200':
          description: Updated profile.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/XRequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CandidateProfile'
        default:
          $ref: '#/components/responses/ApiErrorResponse'
...
  /jd-match/profile:
    get:
      tags: [JobMatch]
      operationId: getJobMatchProfile
      summary: Get the candidate profile snapshot used for JD matching
      description: |
        Returns the aggregated profile snapshot the matching agent uses to score
        recommendations. Aggregated from resumes / target jobs / mocks /
        debriefs by the future `backend-jobs-recommendations` subspec; the v1.0.0
        contract is fixture-backed.
      parameters:
        - $ref: '#/components/parameters/XRequestID'
        - $ref: '#/components/parameters/Traceparent'
        - $ref: '#/components/parameters/AcceptLanguage'
        - $ref: '#/components/parameters/XClientVersion'
      responses:
        '200':
          description: Current profile snapshot.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/XRequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobMatchProfile'
        default:
          $ref: '#/components/responses/ApiErrorResponse'
  /jd-match/agent-status:
    get:
      tags: [JobMatch]
      operationId: getAgentScanStatus
      summary: Get the background JD-match agent scan status
      description: |
        Returns the current agent scan status (`idle` / `scanning` / `error`)
        plus last/next scan timestamps. Frontend polls only on JD-match entry
        and tab returns; no SSE/WebSocket per plan D-10.
      parameters:
        - $ref: '#/components/parameters/XRequestID'
        - $ref: '#/components/parameters/Traceparent'
        - $ref: '#/components/parameters/AcceptLanguage'
        - $ref: '#/components/parameters/XClientVersion'
      responses:
        '200':
          description: Current agent scan status.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/XRequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentScanStatus'
        default:
          $ref: '#/components/responses/ApiErrorResponse'
...
  /jd-match/watchlist:
    get:
      tags: [JobMatch]
      operationId: listWatchlist
      summary: List the candidate watchlist
      description: |
        Returns the candidate watchlist; each item carries `linkedJobMatchId`
        for the chevron handoff back to the Recommended tab.
      parameters:
        - $ref: '#/components/parameters/XRequestID'
        - $ref: '#/components/parameters/Traceparent'
        - $ref: '#/components/parameters/AcceptLanguage'
        - $ref: '#/components/parameters/XClientVersion'
      responses:
        '200':
          description: Watchlist items.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/XRequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WatchlistResponse'
        default:
          $ref: '#/components/responses/ApiErrorResponse'
    post:
      tags: [JobMatch]
      operationId: addToWatchlist
      summary: Save a recommendation to the watchlist
      description: Adds the referenced recommendation to the candidate watchlist.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/XRequestID'
        - $ref: '#/components/parameters/Traceparent'
        - $ref: '#/components/parameters/AcceptLanguage'
        - $ref: '#/components/parameters/XClientVersion'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddToWatchlistRequest'
      responses:
        '200':
          description: Watchlist item created or refreshed.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/XRequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WatchlistItem'
        default:
          $ref: '#/components/responses/ApiErrorResponse'
  /jd-match/watchlist/{jobMatchId}:
    delete:
      tags: [JobMatch]
      operationId: removeFromWatchlist
      summary: Remove a recommendation from the watchlist
      description: Drops the watchlist entry that links to the given recommendation id.
      parameters:
        - name: jobMatchId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/XRequestID'
        - $ref: '#/components/parameters/Traceparent'
        - $ref: '#/components/parameters/AcceptLanguage'
        - $ref: '#/components/parameters/XClientVersion'
      responses:
        '204':
          description: Watchlist entry removed.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/XRequestID'
        default:
          $ref: '#/components/responses/ApiErrorResponse'
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: ei_session
      description: |
        First-party session cookie minted by `GET /api/v1/auth/email/verify`.
        ADR-Q1 locks the cookie attributes (HttpOnly, SameSite=Lax, Secure, ~30d
        sliding renewal). Cookie value is opaque; the server-side `sessions`
        table is the source of truth. `Authorization: Bearer` is **not** an
        accepted P0 scheme — adding it requires a revision of ADR-Q1 and this
        spec.
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: |
        Idempotency key per `B1 shared-conventions-codified §3.4` (24h TTL). Required by
        side-effect endpoints listed in spec D-6, except `POST /auth/email/start`
        (rate-limited per ADR-Q1) and `POST /practice/sessions/{sessionId}/events`
        (uses `clientEventId` instead).
      schema:
        type: string
        format: uuid
    XRequestID:
      name: X-Request-ID
      in: header
      required: false
      description: Client-supplied correlation id; the server echoes it back via the X-Request-ID response header and `error.requestId`.
      schema:
        type: string
    Traceparent:
      name: traceparent
      in: header
      required: false
      description: W3C Trace Context header.
      schema:
        type: string
    AcceptLanguage:
      name: Accept-Language
      in: header
      required: false
      description: |
        UI-locale hint per `B1 shared-conventions-codified §3.5`. Affects display defaults only;
        does **not** override persisted business fields such as `targetLanguage`.
      schema:
        type: string
    XClientVersion:
      name: X-Client-Version
      in: header
      required: false
      description: Client build version reported by the frontend shell.
      schema:
        type: string
  headers:
    XRequestID:
      description: Server-confirmed correlation id; matches `error.requestId` on error responses.
      schema:
        type: string
  responses:
    ApiErrorResponse:
      description: Error envelope per `ApiError` schema.
      headers:
        X-Request-ID:
          $ref: '#/components/headers/XRequestID'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
  schemas:
...
    ApiErrorCode:
      type: string
      description: |
        Documented error codes synced from
        `shared/conventions.yaml#errors[].code`; B1 owns the literal set.
        P0 export exceptions such as `PRIVACY_EXPORT_NOT_AVAILABLE` and
        `RESUME_EXPORT_NOT_AVAILABLE` are included here so the OpenAPI
        contract is self-contained for 501 export responses.
      enum:
        - AUTH_UNAUTHORIZED
        - TARGET_IMPORT_FAILED
        - TARGET_JOB_NOT_FOUND
        - TARGET_IMPORT_SOURCE_INVALID
        - TARGET_IMPORT_SOURCE_UNAVAILABLE
        - TARGET_INVALID_STATE_TRANSITION
        - PRACTICE_SESSION_CONFLICT
        - PRACTICE_PLAN_NOT_FOUND
        - PRACTICE_SESSION_NOT_FOUND
        - REPORT_NOT_FOUND
        - REPORT_NOT_READY
        - DEBRIEF_NOT_FOUND
        - RESUME_EXPORT_NOT_AVAILABLE
        - RESUME_STRUCTURED_MASTER_ALREADY_EXISTS
        - VALIDATION_FAILED
        - RESOURCE_NOT_FOUND
        - IDEMPOTENCY_KEY_MISMATCH
        - RATE_LIMITED
        - AI_PROVIDER_TIMEOUT
        - AI_OUTPUT_INVALID
        - AI_FALLBACK_EXHAUSTED
        - AI_UNSUPPORTED_CAPABILITY
        - AI_PROVIDER_CONFIG_INVALID
        - AI_PROVIDER_SECRET_MISSING
        - PRIVACY_EXPORT_NOT_AVAILABLE
    ApiError:
      type: object
      required: [code, message, requestId, retryable]
      description: |
        Canonical error object per B1 `structures.ApiError`. B2 wraps it in
        `ApiErrorResponse` for the wire body `{error: ...}`.
      properties:
        code:
          $ref: '#/components/schemas/ApiErrorCode'
        message:
          type: string
          description: Human-readable summary; safe to surface to operators (no secrets, per A4 redaction rules).
        requestId:
          type: string
          description: Request correlation id; equals the X-Request-ID response header for the same request.
        retryable:
          type: boolean
          description: Hint to the client whether retry is safe, as defined by B1 shared-conventions-codified.
        details:
          type: object
          description: Optional structured diagnostic. Must not contain provider, prompt, or secret material per A4 redaction rules.
          additionalProperties: true
    ApiErrorResponse:
      type: object
      required: [error]
      description: Wire error response envelope defined by B1 shared-conventions-codified.
      properties:
        error:
          $ref: '#/components/schemas/ApiError'
...
    CandidateProfile:
      type: object
      required: [headline, yearsOfExperience, currentRole, preferredPracticeLanguage, uiLanguage, region]
      properties:
        headline:
          type: string
          nullable: true
        yearsOfExperience:
          type: integer
          format: int32
          minimum: 0
          nullable: true
        currentRole:
          type: string
          nullable: true
        preferredPracticeLanguage:
          type: string
        uiLanguage:
          type: string
        region:
          type: string
          nullable: true
    UpdateProfileRequest:
      type: object
      description: All fields optional — only supplied fields are updated.
      properties:
        headline:
          type: string
        yearsOfExperience:
          type: integer
          format: int32
          minimum: 0
        currentRole:
          type: string
        preferredPracticeLanguage:
          type: string
        uiLanguage:
          type: string
        region:
          type: string
...
    JobMatchProfileSourceCounts:
      type: object
      required: [resumes, jds, mocks, debriefs]
      description: |
        Aggregated counts of inputs that fed the JD-match profile snapshot.
        Driven by the future `backend-jobs-recommendations` subspec.
      properties:
        resumes:
          type: integer
          format: int32
          minimum: 0
        jds:
          type: integer
          format: int32
          minimum: 0
        mocks:
          type: integer
          format: int32
          minimum: 0
        debriefs:
          type: integer
          format: int32
          minimum: 0
    JobMatchProfile:
      type: object
      required: [displayName, skills, sources]
      description: |
        Candidate profile snapshot used by the JD-match agent. Returned by
        `getJobMatchProfile`. Aggregated server-side; the frontend never derives
        the snapshot from individual resumes.
      properties:
        displayName:
          type: string
          description: Candidate display name; never the raw email.
        avatarUrl:
          oneOf:
            - type: string
              format: uri
            - type: 'null'
        headline:
          oneOf:
            - type: string
            - type: 'null'
        yearsOfExperience:
          oneOf:
            - type: integer
              format: int32
              minimum: 0
            - type: 'null'
        locationText:
          oneOf:
            - type: string
            - type: 'null'
        compensationText:
          oneOf:
            - type: string
            - type: 'null'
        skills:
          type: array
          items:
            type: string
        sources:
          $ref: '#/components/schemas/JobMatchProfileSourceCounts'
    JobMatchAgentStatus:
      type: string
      description: |
        Agent scan status. `idle` = no scan running; `scanning` = scan in
        progress; `error` = last scan failed and operator action is needed.
      enum: [idle, scanning, error]
    AgentScanStatus:
      type: object
      required: [status]
      description: |
        Background JD-match agent status returned by `getAgentScanStatus`.
        Frontend reads only on JD-match entry and tab returns; no SSE/WebSocket
        per plan D-10.
      properties:
        status:
          $ref: '#/components/schemas/JobMatchAgentStatus'
        lastScanAt:
          oneOf:
            - type: string
              format: date-time
            - type: 'null'
        nextScanAt:
          oneOf:
            - type: string
              format: date-time
            - type: 'null'
        message:
          oneOf:
            - type: string
            - type: 'null'
          description: Optional human-readable hint (only used when status == error).
...
    AddToWatchlistRequest:
      type: object
      required: [jobMatchId]
      properties:
        jobMatchId:
          type: string
          format: uuid
    WatchlistItemTone:
      type: string
      description: Visual tone driving watchlist row colour per ui-design tone tokens.
      enum: [ok, warn, muted]
    WatchlistItem:
      type: object
      required: [id, linkedJobMatchId, title, company, tone, addedAt]
      description: |
        A single watchlist row. `linkedJobMatchId` powers the chevron handoff to
        the Recommended tab; the frontend must read it from the schema rather
        than string-matching ids.
      properties:
        id:
          type: string
          format: uuid
        linkedJobMatchId:
          type: string
          format: uuid
        label:
          oneOf:
            - type: string
            - type: 'null'
          description: Optional candidate-supplied label; never logged.
        title:
          type: string
        company:
          type: string
        tone:
          $ref: '#/components/schemas/WatchlistItemTone'
        addedAt:
          type: string
          format: date-time
        change:
          oneOf:
            - type: string
            - type: 'null'
          description: Backend-rendered change signal (e.g., "JD updated · check diff").
    WatchlistResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/WatchlistItem'
...